Extract performance evidence view builder
Move performance evidence-view request context, fetch state, and response-state resolution into the evidence helper module while preserving gateway behavior.

Evidence:

- python -m ruff format --check touched service/test files: passed

- python -m ruff check touched service/test files: passed

- python -m mypy src/app/services/performance_workspace_service.py src/app/services/performance_workspace_evidence.py: passed

- python -m pytest tests/unit/test_performance_workspace_evidence.py tests/unit/test_performance_workspace_service.py: 51 passed

- make check: 1,063 unit/contract tests passed

- make ci: 207 integration tests and 1,270 coverage tests passed, total coverage 94.05%, pip-audit no known vulnerabilities with governed PYSEC-2026-161 exception
