+++ quality/quality_scorecard.md
@@ -5,21 +5,21 @@ Mode: feature-branch evidence refresh
 
 | Dimension | Score | Current status | Next action |
 | --- | ---: | --- | --- |
-| Build and test reliability | 5/5 | Current performance evidence contract branch `make check` passed with ruff, format check, monetary-float guard, mypy over 471 source files, Workbench/OpenAPI contract smoke, and 1,059 unit/contract tests; `make ci` passed with 207 integration tests, 1,266 coverage tests, 94.05% total coverage, and dependency audit | Keep Docker parity blocking in PR Merge Gate |
+| Build and test reliability | 5/5 | Current performance evidence-view builder branch `make check` passed with ruff, format check, monetary-float guard, mypy over 471 source files, Workbench/OpenAPI contract smoke, and 1,063 unit/contract tests; `make ci` passed with 207 integration tests, 1,270 coverage tests, 94.05% total coverage, and dependency audit | Keep Docker parity blocking in PR Merge Gate |
 | Coverage | 4/5 | 94.05% total coverage, above the 84% floor | Add targeted middleware/security/error tests |
-| Modularity | 4/5 | Current branch state preserves the 49-line longest-function baseline; recent slices extracted transaction request context, transaction page-context defaults, transaction client-kwargs mapping, performance workspace response assembly, portfolio workspace response assembly, portfolio workspace performance parsing, portfolio workspace rebalance parsing, portfolio source-readiness parsing, portfolio transaction summary mapping, portfolio workflow mapping, portfolio workflow contracts, portfolio transaction contracts, portfolio performance snapshot contracts, portfolio income/activity contracts, portfolio holdings/book contracts, risk drawdown contracts, reporting batch contracts, reporting query contracts, risk concentration contracts, risk rolling contracts, risk attribution contracts, performance contribution contracts, performance attribution contracts, and performance evidence contracts; `portfolio_service.py` is now 1,970 measured lines, `performance_workspace.py` is now 903 measured lines, `portfolio.py` is now 911 measured lines, `risk_workspace.py` is now 678 measured lines, and `reporting.py` is now 532 measured lines | Continue extraction slices and reduce remaining largest-file pressure |
+| Modularity | 4/5 | Current branch state preserves the 49-line longest-function baseline; recent slices extracted transaction request context, transaction page-context defaults, transaction client-kwargs mapping, performance workspace response assembly, portfolio workspace response assembly, portfolio workspace performance parsing, portfolio workspace rebalance parsing, portfolio source-readiness parsing, portfolio transaction summary mapping, portfolio workflow mapping, portfolio workflow contracts, portfolio transaction contracts, portfolio performance snapshot contracts, portfolio income/activity contracts, portfolio holdings/book contracts, risk drawdown contracts, reporting batch contracts, reporting query contracts, risk concentration contracts, risk rolling contracts, risk attribution contracts, performance contribution contracts, performance attribution contracts, performance evidence contracts, and performance evidence-view orchestration; `portfolio_service.py` is now 1,970 measured lines, `performance_workspace_service.py` is now 1,413 measured lines, `performance_workspace.py` is now 903 measured lines, `portfolio.py` is now 911 measured lines, `risk_workspace.py` is now 678 measured lines, and `reporting.py` is now 532 measured lines | Continue extraction slices and reduce remaining largest-file pressure |
 | Architecture boundaries | 3/5 | Blocking AST tests exist; import-linter is report-only | Classify and enforce no-new-regression |
 | API governance | 4/5 | 233 OpenAPI paths and 247 operations; missing summary, description, operation ID, tags, and documented 4xx/5xx response counts are all 0; Spectral remains report-only | Triage Spectral warnings and decide explicit operation ID policy |
 | Error consistency | 2/5 | ProblemDetails exists for unhandled exceptions; route/upstream error normalization remains a meaningful hardening candidate | Normalize route/upstream errors |
 | Observability | 3/5 | Health/readiness/metrics/correlation/audit are present; RFC-0108 fan-out and selected analytics audit posture remain implementation-backed | Enforce structured log and metric label rules |
-| Security | 4/5 | Latest merged performance attribution slice `pip-audit` passed with the governed FastAPI/Starlette exception; current performance evidence contract branch monetary-float guard passed without allowlist churn | Triage bandit and sensitive-data handling checks |
+| Security | 4/5 | Current performance evidence-view builder branch `pip-audit` passed with the governed FastAPI/Starlette exception; monetary-float guard passed without allowlist churn | Triage bandit and sensitive-data handling checks |
 | Documentation | 4/5 | Baseline, scorecard, health report, and wiki validation evidence are refreshed to current branch metrics after the current focused performance attribution contract slice | Keep wiki synced and add diagrams over time |
 | Operations readiness | 3/5 | Existing CI/runbook docs and wiki validation posture describe the report-only quality lane | Add incident playbooks and SLO checks |
 
 ## Before/After Evidence
 
 Comparison point: the prior scorecard state after the portfolio liquidity payload-loader slice
-versus the current performance evidence contract branch after PRs #349, #350, #351, #352,
+versus the current performance evidence-view builder branch after PRs #349, #350, #351, #352,
 #353, #354, #355, #356, #357, #358, #359, #360, #361, #362, #363, #364, #365, #366, #367,
 #368, #369, #370, #371, and #372.
 
@@ -33,8 +33,8 @@ versus the current performance evidence contract branch after PRs #349, #350, #3
 | Largest source file | 2,968 lines | 1,970 lines | Improved; `src/app/services/portfolio_service.py` remains the largest residual hotspot after reducing `performance_workspace.py` to 903 lines, `portfolio.py` to 911 lines, `risk_workspace.py` to 678 lines, and `reporting.py` to 532 lines |
 | `performance_workspace.py` | 1,539 lines | 903 lines | Improved through performance contribution, attribution, and evidence contract extraction |
 | OpenAPI operations with missing summary/description/tags/errors | 0 | 0 | Preserved |
-| Local unit/contract tests | 996 | 1,059 | Added focused response/request boundary, workspace parser, source-readiness parser, transaction summary mapper, workflow mapper, workflow contract, transaction contract, performance snapshot contract, income/activity contract, holdings contract, risk drawdown contract, reporting batch contract, reporting query contract, risk concentration contract, risk rolling contract, risk attribution contract, performance contribution contract, performance attribution contract, and performance evidence compatibility tests |
-| Local coverage tests | 1,203 | 1,266 | Added focused coverage while preserving total coverage |
+| Local unit/contract tests | 996 | 1,063 | Added focused response/request boundary, workspace parser, source-readiness parser, transaction summary mapper, workflow mapper, workflow contract, transaction contract, performance snapshot contract, income/activity contract, holdings contract, risk drawdown contract, reporting batch contract, reporting query contract, risk concentration contract, risk rolling contract, risk attribution contract, performance contribution contract, performance attribution contract, performance evidence compatibility tests, and performance evidence-view builder tests |
+| Local coverage tests | 1,203 | 1,270 | Added focused coverage while preserving total coverage |
 | Total coverage | 93.69% | 94.05% | Improved |
 | Dependency audit | governed pass | governed pass, no known vulnerabilities after the `PYSEC-2026-161` exception | Preserved |
