+++ quality/refactor_health_report.md
@@ -322,17 +322,23 @@ legacy `app.contracts.performance_workspace` import surface intact, preserves th
 monetary-float allowlist without churn, and reduces `performance_workspace.py` from 1,101 to 903
 lines while preserving focused evidence/capabilities/response tests and the 49-line
 longest-function baseline.
+The current performance evidence-view builder slice moves evidence request context, fetch state,
+source-supportability collection, durable calculation evidence fetching, and supported/partial/
+unavailable evidence response resolution into `performance_workspace_evidence.py`, reducing
+`performance_workspace_service.py` from 1,611 to 1,413 measured lines while preserving focused
+evidence/service tests, Workbench performance evidence behavior, and the 49-line longest-function
+baseline.
 
 ## Health Signals
 
 | Area | Current posture | Evidence |
 | --- | --- | --- |
-| Branch hygiene | Healthy | Current performance evidence contract branch was created from clean `main` after PR #372; final remote/local cleanup remains a post-merge gate |
-| Unit/contract coverage | Healthy | Current performance evidence contract branch `make check` passed with ruff, format check, monetary-float guard, mypy over 471 source files, Workbench/OpenAPI contract smoke, and 1,059 unit/contract tests |
-| Integration coverage | Healthy | 207 integration tests passed in current performance evidence contract branch `make ci` |
-| Total coverage | Healthy | 1,266 coverage tests passed in current performance evidence contract branch `make ci`; total coverage is 94.05%, above the 84% floor |
-| Security audit | Governed | Current performance evidence contract branch `pip-audit` found no known vulnerabilities after the governed `PYSEC-2026-161` exception; monetary-float guard passed without allowlist churn |
-| Modularity | Improving, incomplete | Longest-function baseline remains 49 lines; recent response/request-context/parser/mapper/contract slices reduce `performance_workspace.py` to 903 measured lines, `portfolio.py` to 911 measured lines, `risk_workspace.py` to 678 measured lines, and `reporting.py` to 532 measured lines, leaving `portfolio_service.py` at 1,970 measured lines and `performance_workspace_service.py` at 1,611 measured lines; several service files remain above 1,000 lines |
+| Branch hygiene | Healthy | Current performance evidence-view builder branch was created from clean `main` after PR #373; final remote/local cleanup remains a post-merge gate |
+| Unit/contract coverage | Healthy | Current performance evidence-view builder branch `make check` passed with ruff, format check, monetary-float guard, mypy over 471 source files, Workbench/OpenAPI contract smoke, and 1,063 unit/contract tests |
+| Integration coverage | Healthy | 207 integration tests passed in current performance evidence-view builder branch `make ci` |
+| Total coverage | Healthy | 1,270 coverage tests passed in current performance evidence-view builder branch `make ci`; total coverage is 94.05%, above the 84% floor |
+| Security audit | Governed | Current performance evidence-view builder branch `pip-audit` found no known vulnerabilities after the governed `PYSEC-2026-161` exception; monetary-float guard passed without allowlist churn |
+| Modularity | Improving, incomplete | Longest-function baseline remains 49 lines; recent response/request-context/parser/mapper/contract slices reduce `performance_workspace_service.py` to 1,413 measured lines, `performance_workspace.py` to 903 measured lines, `portfolio.py` to 911 measured lines, `risk_workspace.py` to 678 measured lines, and `reporting.py` to 532 measured lines, leaving `portfolio_service.py` at 1,970 measured lines and `advisor_brief_service.py` at 1,454 measured lines; several service files remain above 1,000 lines |
 | API governance | Improving, incomplete | 233 OpenAPI paths and 247 operations have summaries, descriptions, operation IDs, tags, and documented 4xx/5xx responses; Spectral remains report-only |
 | Architecture rules | Improving, incomplete | AST boundary tests exist; import-linter is new report-only baseline |
 | Observability | Partial | Health/readiness/metrics/correlation exist; trace/log scoring not enforced |
